Anagrafica prodotti
Client requested that products are to be client specific.
Each client has its own products with prices. Each product can be
present many times in the listing with different starting date.

This commit modifies the DB structure and object model to comply to the
request. UI and BL are still to be refactored.