ViaggioEdit Refactoring
Moved some code from ViaggioEditVModel to ViaggioEditPresenter since it
belonged to business logic and not to presentation.

A big issue is still present.
Since the client requested Calo Peso to be unique for the Viaggio, all
the rows whose weight is less than the Calo Peso will have a negative
cost!

This issue has to be brought to client's attention ... again.