ViaggioEditView Edit Viaggio Checks
While editing a Viaggio, if the user picks a date for which there are no
active products, a message is shown to inform the user and the Viaggio's
rows are cleared.

If a CaloPeso is specified which leads to a <=0 cost for any row, a
message is shown to inform the user.